requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.13
django-bootstrap-v5==1.0.11
django-ckeditor==6.6.1
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==22.1
django-heroku==0.3.1
django-js-asset==2.1.0
django-tinymce==3.5.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
gunicorn==22.0.0
idna==3.7
imageio==2.34.1
joblib==1.4.2
lazy_loader==0.4
networkx==3.2.1
numpy==1.24.3
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tifffile==2024.5.22
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
